requirements.txt
PyMuPDF>=1.23.0
Pillow>=10.0.0
customtkinter>=5.2.0
CTkMessagebox>=2.0.0
